rename specific.c -> proc.c
