+++ MANIFEST.in
@@ -55,12 +55,12 @@ include psutil/arch/freebsd/disk.c
 include psutil/arch/freebsd/disk.h
 include psutil/arch/freebsd/mem.c
 include psutil/arch/freebsd/mem.h
+include psutil/arch/freebsd/proc.c
+include psutil/arch/freebsd/proc.h
 include psutil/arch/freebsd/proc_socks.c
 include psutil/arch/freebsd/proc_socks.h
 include psutil/arch/freebsd/sensors.c
 include psutil/arch/freebsd/sensors.h
-include psutil/arch/freebsd/specific.c
-include psutil/arch/freebsd/specific.h
 include psutil/arch/freebsd/sys_socks.c
 include psutil/arch/freebsd/sys_socks.h
 include psutil/arch/netbsd/socks.c
