+++ psutil/_psutil_bsd.c
@@ -67,7 +67,7 @@
     #include "arch/freebsd/mem.h"
     #include "arch/freebsd/disk.h"
     #include "arch/freebsd/sensors.h"
-    #include "arch/freebsd/specific.h"
+    #include "arch/freebsd/proc.h"
     #include "arch/freebsd/sys_socks.h"
     #include "arch/freebsd/proc_socks.h"
