+++ setup.py
@@ -210,7 +210,7 @@ elif FREEBSD:
             'psutil/arch/freebsd/mem.c',
             'psutil/arch/freebsd/disk.c',
             'psutil/arch/freebsd/sensors.c',
-            'psutil/arch/freebsd/specific.c',
+            'psutil/arch/freebsd/proc.c',
             'psutil/arch/freebsd/sys_socks.c',
             'psutil/arch/freebsd/proc_socks.c',
         ],
